Adjust package directory of sharding algorithm
