+++ sharding-core/sharding-core-common/src/main/java/org/apache/shardingsphere/core/strategy/algorithm/sharding/inline/InlineExpressionParser.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.underlying.common.config.inline;
+package org.apache.shardingsphere.core.strategy.algorithm.sharding.inline;
 
 import com.google.common.base.Joiner;
 import com.google.common.collect.Collections2;
