+++ sharding-core/sharding-core-common/src/test/java/org/apache/shardingsphere/core/strategy/algorithm/sharding/inline/InlineExpressionParserTest.java
@@ -15,8 +15,9 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.underlying.common.config.inline;
+package org.apache.shardingsphere.core.strategy.algorithm.sharding.inline;
 
+import org.apache.shardingsphere.core.strategy.algorithm.sharding.inline.InlineExpressionParser;
 import org.junit.Test;
 
 import java.util.Collections;
