+++ sharding-integration-test/sharding-jdbc-test/src/test/java/org/apache/shardingsphere/dbtest/cases/dataset/DataSet.java
@@ -19,7 +19,7 @@ package org.apache.shardingsphere.dbtest.cases.dataset;
 
 import lombok.Getter;
 import org.apache.shardingsphere.underlying.common.rule.DataNode;
-import org.apache.shardingsphere.underlying.common.config.inline.InlineExpressionParser;
+import org.apache.shardingsphere.core.strategy.algorithm.sharding.inline.InlineExpressionParser;
 import org.apache.shardingsphere.dbtest.cases.dataset.metadata.DataSetMetadata;
 import org.apache.shardingsphere.dbtest.cases.dataset.row.DataSetRow;
