+++ sharding-jdbc/sharding-jdbc-core/src/test/resources/META-INF/services/org.apache.shardingsphere.spi.keygen.KeyGenerateAlgorithm
@@ -15,6 +15,6 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.core.strategy.keygen.SnowflakeKeyGenerateAlgorithm
-org.apache.shardingsphere.core.strategy.keygen.UUIDKeyGenerateAlgorithm
+org.apache.shardingsphere.core.strategy.algorithm.keygen.SnowflakeKeyGenerateAlgorithm
+org.apache.shardingsphere.core.strategy.algorithm.keygen.UUIDKeyGenerateAlgorithm
 org.apache.shardingsphere.shardingjdbc.fixture.IncrementKeyGenerateAlgorithm
