+++ sharding-jdbc/sharding-jdbc-orchestration/src/test/resources/META-INF/services/org.apache.shardingsphere.spi.keygen.KeyGenerateAlgorithm
@@ -15,7 +15,7 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.core.strategy.keygen.SnowflakeKeyGenerateAlgorithm
-org.apache.shardingsphere.core.strategy.keygen.UUIDKeyGenerateAlgorithm
+org.apache.shardingsphere.core.strategy.algorithm.keygen.SnowflakeKeyGenerateAlgorithm
+org.apache.shardingsphere.core.strategy.algorithm.keygen.UUIDKeyGenerateAlgorithm
 org.apache.shardingsphere.shardingjdbc.orchestration.api.yaml.fixture.DecrementKeyGenerateAlgorithm
 org.apache.shardingsphere.shardingjdbc.orchestration.api.yaml.fixture.IncrementKeyGenerateAlgorithm
