+++ sharding-orchestration/sharding-orchestration-core/sharding-orchestration-core-common/src/test/java/org/apache/shardingsphere/orchestration/core/common/rule/OrchestrationShardingRuleTest.java
@@ -23,7 +23,7 @@ import org.apache.shardingsphere.api.config.sharding.ShardingRuleConfiguration;
 import org.apache.shardingsphere.api.config.sharding.TableRuleConfiguration;
 import org.apache.shardingsphere.api.config.sharding.strategy.NoneShardingStrategyConfiguration;
 import org.apache.shardingsphere.api.config.sharding.strategy.StandardShardingStrategyConfiguration;
-import org.apache.shardingsphere.core.strategy.sharding.InlineShardingAlgorithm;
+import org.apache.shardingsphere.core.strategy.algorithm.sharding.inline.InlineShardingAlgorithm;
 import org.junit.Before;
 import org.junit.Test;
